% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sift.R
\name{sift}
\alias{sift}
\title{Extract & summarise function usage}
\usage{
sift(fil)
}
\arguments{
\item{fil}{A file name (.R, .Rmd or .qmd)}
}
\value{
A tibble
}
\description{
Extract & summarise function usage
}
\examples{
# Mimics the input of a two-line R script
sift("mean(c(1, 2, 3))\nsum(c(1, 2, 3))")
#   A tibble: 1 × 2
#   Package Function
#   <chr>   <chr>
# 1 base    c[2];  mean[1];  sum[1]



}
