% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usedthese-package.R
\docType{package}
\name{usedthese-package}
\alias{usedthese}
\alias{usedthese-package}
\title{usedthese: Summarises Package & Function Usage}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Consistent with 'knitr' syntax highlighting, 'usedthese' adds a summary table of package & function usage to a Quarto document and enables aggregation of usage across a website.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cgoo4.github.io/usedthese/}
  \item \url{https://github.com/cgoo4/usedthese}
  \item Report bugs at \url{https://github.com/cgoo4/usedthese/issues}
}

}
\author{
\strong{Maintainer}: Carl Goodwin \email{carl.goodwin@quantumjitter.com} [copyright holder]

Authors:
\itemize{
  \item Carl Goodwin \email{carl.goodwin@quantumjitter.com} [copyright holder]
}

}
\keyword{internal}
