% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sift.R
\name{sift}
\alias{sift}
\title{Extract & summarise function usage}
\usage{
sift()
}
\arguments{
\item{x}{A file name -- .R, .Rmd or .qmd}
}
\value{
Printed table
}
\description{
Extract & summarise function usage
}
\examples{
xxx

}
