% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/used_here.R
\name{used_here}
\alias{used_here}
\title{Summarise function usage here}
\usage{
used_here(fil = knitr::current_input())
}
\arguments{
\item{fil}{If the usage summary is required in the document you are currently knitting the current document, it is not necessary to specify
then no argument need be specified.

If you want to create the summary by running the code chunk directly, then it is necessary to
specify the quoted name of the saved file. This would also be the case if the usage summary
is required in, say, a blog post that's summarising the code in some other file.}
}
\value{
A printed kable table with the css class "usedthese"
}
\description{
Consistent with knitr syntax highlighting, [used_here()] adds a
summary table of R package & function usage to a knitted Quarto or R Markdown document
}
\details{
If the rendered summary includes rows where the package name is prefixed with "DUPE", this will
be due to a conflict [used_here()] was unable to resolve, e.g. [dplyr::first()] versus
[xts::first()]. In such cases, the recommendation is to add "exclude = first" to the library
call of the unwanted version.
}
\examples{
# Simple example -- see Get Started for an extended example
mean(c(1, 2, 3))
sum(c(1, 2, 3))

# used_here()
# Package Function
# base    c[2];  mean[1];  sum[1]

}
