% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/used_here.R
\name{used_here}
\alias{used_here}
\title{Extract & summarise function usage}
\usage{
used_here(fil = knitr::current_input())
}
\arguments{
\item{fil}{A file name (.R, .Rmd or .qmd)}
}
\value{
A tibble
}
\description{
Extract & summarise function usage
}
\examples{
# Mimics the input of a two-line R script
used_here("mean(c(1, 2, 3))\nsum(c(1, 2, 3))")
#   A tibble: 1 × 2
#   Package Function
#   <chr>   <chr>
# 1 base    c[2];  mean[1];  sum[1]

}
